+++ foundation/memory.py
@@ -1,3 +1,20 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+#   Copyright 2015 Futur Solo
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
 
 from tornado.gen import *
 from collections import OrderedDict
@@ -61,6 +78,13 @@ class Element:
             update = {"$set": update_condition}
         self._action = self._current_collection.find_and_modify(
             query=condition, update=update, new=new)
+        self._action_ready = True
+        return self
+
+    def count(self):
+        self.reset()
+        self._action = self._current_collection.count()
+        self._action_ready = True
         return self
 
     def skip(self, skip=0):