+++ spirit/management/framework.js
@@ -12,3 +12,19 @@
 //   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
+
+$(".textarea textarea").keypress(function(){
+    $(this).parent(".textarea").children("pre").html($(this).val());
+});
+$(".textarea textarea").keydown(function(){
+    $(this).parent(".textarea").children("pre").html($(this).val());
+});
+$(".textarea textarea").keyup(function(){
+    $(this).parent(".textarea").children("pre").html($(this).val());
+});
+$(".textarea textarea").change(function(){
+    $(this).parent(".textarea").children("pre").html($(this).val());
+});
+$(".textarea textarea").blur(function(){
+    $(this).parent(".textarea").children("pre").html($(this).val());
+});