+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+v20150401 dev
+-------------
+ - Management Enhance
+ - Changed Template system to Mako
+ - Completed Apache License Comments
+ - Bugs Fix
+ - Font Family Update
+ - Footer Added
+
+
 v20150331 dev
 -------------
  - Bugs Fix